Added a script for pulling a map of usernames and hashed emails
To be able to merge data between the open discussions application and the mailgun logs we need a mapping of user ids and hashed email addresses. This script generates a parquet file of that mapping and uploasds it to S3 for being picked up by AWS glue and exposed to AWS Athena.